docs(skill): make chat usage helper runtime-agnostic
Replace the OpenClaw extractor with chat_usage.py: normalize hop totals
you already have. Vendor trees stay out of the skill script.
